Checkpoint current changes
