+++ .env.example
@@ -11,8 +11,10 @@ EMBEDDING_MODEL="nomic-embed-text"
 # Embedding vector dimension (required, must match the model above)
 EMBEDDING_DIMENSION="768"
 
-# Optional chunk-size override (otherwise derived from the embedding model)
+# Optional chunking overrides. Leave commented to use the automatic heuristic (window/4, clamped to 256-1024).
 # TEXT_SPLITTER_CHUNK_SIZE="1024"
+# TEXT_SPLITTER_CHUNK_OVERLAP="64"
+# TEXT_SPLITTER_USE_SAFE_DEFAULTS="1"
 
 # Optional search tuning knobs
 # SEARCH_DEFAULT_LIMIT="5"
