+++ Cargo.toml
@@ -52,6 +52,7 @@ cargo-husky = { version = "1", default-features = false, features = [
 ] }
 httpmock = "0.7"
 regex = "1"
+tower = { version = "0.5", features = ["util"] }
 
 [package.metadata.cargo-machete]
 ignored = ["cargo-husky"]
