+++ src/logging.rs
@@ -1,3 +1,9 @@
+//! Tracing configuration and log routing.
+//!
+//! The application logs to stdout using a compact formatter, and optionally to a file. When
+//! `RUSTY_MEM_LOG_FILE` is set, logs are appended to that path; otherwise a file logger is
+//! created under `logs/rusty-mem.log`. A non‑blocking writer is used to minimize contention
+//! on hot paths.
 use std::sync::OnceLock;
 
 use tracing_appender::non_blocking::{NonBlocking, WorkerGuard};
@@ -6,6 +12,10 @@ use tracing_subscriber::{EnvFilter, fmt, prelude::*};
 static LOG_GUARD: OnceLock<WorkerGuard> = OnceLock::new();
 
 /// Configure tracing subscribers for stdout and optional file logging.
+///
+/// - Respects `RUST_LOG` for filtering (defaults to `info`).
+/// - Installs a compact stdout layer and, when available, a file layer.
+/// - Uses a global guard to keep the non‑blocking writer alive for the process lifetime.
 pub fn init_tracing() {
     let env_filter = EnvFilter::try_from_default_env().unwrap_or_else(|_| EnvFilter::new("info"));
     let stdout_layer = fmt::layer().with_target(false).compact();
@@ -27,6 +37,9 @@ pub fn init_tracing() {
     }
 }
 
+/// Build a non‑blocking writer for file logging.
+///
+/// Returns `None` when the logs directory cannot be created or the target file cannot be opened.
 fn configure_file_writer() -> Option<NonBlocking> {
     if let Ok(path) = std::env::var("RUSTY_MEM_LOG_FILE") {
         match std::fs::OpenOptions::new()
